Removed access to twitch API in user factory stuff
Because, you know, it's not required.

Also update user when it leaves.

+++ DamageLib/Database/IConnectionManager.cs
@@ -16,7 +16,7 @@ namespace DamageBot.Database {
         /// </summary>
         /// <param name="command"></param>
         /// <returns></returns>
-        int Write(DbCommand command);
+        long Write(DbCommand command);
 
         /// <summary>
         /// Begins a transaction.
